Create The Leader's Code for a Following. (Not Finallized...)
This is a plan, but I need to add some stuff in this, I will try to work on this in my spare time, but due to scheduling and being distracted with other things, I might not update In a while. I also need to give credit to certain pieces of this code, to the origional makers.